[Major] Fixed bug on extraction (no negative impact)
In the edge case where a low PP variant was at the end
of a FIFO the whole FIFO would be extracted instead of
the flanking (before) hets only. This resulted in more
het variants being extracted. This has no negative
impact since more variants were extracted, so more
info to work with at later stages, however, this was
not the intended behavior. So it was fixed.